Lobby

+++ app/controllers/games_controller.rb
@@ -7,18 +7,26 @@ class GamesController < ApplicationController
 
   def create
     @game = current_user.games.create(game_params)
+    @game.assign_host(current_user)
     @game.make_pieces
+    @game.create_lobby_token(current_user)
     redirect_to game_path(@game)
   end
 
   def show
     @game = Game.find(params[:id])
+    @game.assign_guest(current_user)
+    @game.manage_token
+    SendLobbyTokensJob.perform_later
+    SendGameAndPiecesJob.perform_later(@game)
   end
 
   def destroy
-    game = Game.find(params[:id])
-    game.pieces.destroy_all
-    game.destroy
+    game = Game.find_by_id(params[:id])
+    game.pieces.destroy_all if game
+    game.destroy if game
+    game.lobby_tokens.destroy_all if game
+    SendLobbyTokensJob.perform_later
     redirect_to root_path
   end
